puppet git: add a descriptive config version
The config version is printed aft5er every puppet run, stored in the
state yaml file and can be exported with the prometheous metrics.  This
allow users and automated tooling to easily see what the last change
running on a host is.

The current message is formated as below (but happy to change)
  "$sha1: $author: $message!
e.g.
  d55c65262ad87987cc2fce92f5716000542298da:
  Elukey: role::analytics_cluster::hadoop::worker: deploy TLS keys

Bug: T228854
Change-Id: I8ab540c1521173173b4ce5bf151b4a80f23beb50