gerrit: allow rsync of home dirs for server migrations
Admin users on Gerrit servers tend to have stuff in their home
dirs that should be saved during server migrations.

Add an rsync module to the migration class to allow syncing
home dirs but not directly /home/ to /home/ but into /srv.

Temp. set the "master_host" to cobalt again to allow
it to push to gerrit1001.

Bug: T236187
Bug: T222391
Change-Id: I3f33c8938f06ee38292d8c28136c591a9456922d