vcl: focus on frontends for X-Cache-Status miss and pass
The current logic used to summarize X-Cache into X-Cache-Status
distinguishes between front, local, and remote cache hits. The same goes
for synthetic responses (int).

When it comes to miss and pass, instead, we do not differentiate: if
there's a "miss" anywhere in X-Cache, we count the transaction as a
miss. If not, and there's "pass" anywhere in X-Cache, we count that as a
pass. For this reason, the following X-Cache header is incorrectly
regarded as a miss:

  X-Cache: cp4029 miss, cp4027 pass

This patch restricts the match to the last part (frontend) of X-Cache
for miss and pass.

Change-Id: I2accc4d69dd435a73683fef1d4aee19487fbb4bb