eventlogging: allow sanitization script to run on all db records
The Analytycs team needs to sanitize all the records currently
stored in the eventlogging's databases. By default, the sanitization
script only allows to sanitize records older than 90.
This change allows to run the script with values up to zero, to
enable complete sanitization of all records if needed.

The relative_ts function seems to work fine with zero:

>>> (datetime.utcnow() - timedelta(days=0)).strftime('%Y%m%d%H%M%S')
'20191104085540'

Bug: T236818
Change-Id: I3caa4ec7c964dc70231eec82b2112a73a6ed350f