Bump system user UID range in enforce-users-groups.sh
The new Hiera flag to manage adduser.conf was applied to role(spare), which
reduces LAST_SYSTEM_UID to 499. This leads to the removal of all system users
which have an existing UID in the 500-999 range, but we do have a number of
system users in that range, which need to be cleaned up (e.g. systemd-coredump
and debmonitor).

For now adjust the check.

Bug: T235162
Change-Id: I3721e7d8b0b886da2dd5fa93ddf9fb4a17fb1551