Add analytics users (without ssh keys) to all Hadoop worker nodes
This is a step required before enabling Kerberos authentication on
the hadoop cluster. A required step is in fact setting Yarn to use
proper linux containers on Hadoop workers nodes, launching them
as the user who submitted the job. The current behavior is to launch
jvms directly as the 'yarn' user, hence without the need to have
any user deployed on the hosts.

Bug: T237269
Change-Id: Icba57a8ab975c29cd66cbe9b9aa664034108cde1