puppet-merge: show who is holding the lock
On a failure to acquire the flock, ask fuser which pids are holding
the lock, then run pstree on one of them, with its option enabled to
emit usernames whenever they change along the tree.  This produces
a single line of extra output, is fairly informative, and was fairly
easy to implement (fuser's best efforts aside).

Sample output: https://phabricator.wikimedia.org/P9527

Change-Id: I8733429f0bc509411ec41ed54b65d76921159834