Add Bacula backups for Analytics Meta's mylvm snapshots
Background:
The Analytics team uses mariadb::mylvmbackup on an-coord1001
(where the Analytics Meta mariadb instance runs) to periodically
rsync a lvm snapshot of all the databases to an-master1002.
This is far from ideal, and previously we tried to use proper
mariadb backups for an-coord1001 but we failed, since the table
lock contention, even for low traffic dbs, caused some service
disruptions.

Current need:
We need to have a history of lvm snapshots to be able to recover
data that is not the most recent version of the database. We currently
rely on a systemd timer that copies files to Hadoop's HDFS, but it
turned out to be a hassle with Kerberos (due to permissions etc..).
The ideal solution for the short/medium term is to have bacula to backup
lvm snapshots from an-master1002, to avoid using HDFS for this.

elukey@an-master1002:~$ sudo du -hs /srv/backup/mysql/analytics-meta
18G	/srv/backup/mysql/analytics-meta

Long term plan:
In T234826 the Analytics team is working with Data Persistence to
repurpose db1108 (current Eventlogging database replica) to a generic
Mariadb slave for all Analytics dbs. This will allow us to add Bacula
Backups to it, that will not interfere with the master instances.
When db1108 will be ready this change will be reverted, but a lot of work
is still needed.

Bug: T231208
Change-Id: Ia5846da9d38be8261d5d4db31608f7d728f95bf2