toolforge: Reload/restart services as appropriate on certificate change
Otherwise acme-chief will renew a cert and Puppet on the hosts using the cert
will pull the new one, but won't trigger the existing services to reload it.

For nginx, use the standard reload mechanism. For the k8s master, restart
kube-apiserver per
https://wikitech.wikimedia.org/w/index.php?title=Portal:Cloud_VPS/Admin/SSL_certificate&diff=prev&oldid=1843872

I realised we did this properly in I73841e92 but not Ib7f9a52c.

Change-Id: Ie372bbc8f9dd2e2b2c6d93a62df66f58d53578e5