toolforge: proxy: make https fully optional
The dynamicproxy class supports https being optional, but the profile didn't.
This patch introduces the changes needed to make it hiera-optional.

The main purpose of this is to be able to run this proxy in toolsbeta without
SSL (we don't have the toolforge SSL keys there). The config should be simple,
using this hiera:

  profile::toolforge::proxy::do_https: false

Since the default is true, no hiera changes are required in tools.

Bug: T237443
Change-Id: I6c34f4ddb99be5bd36eb90fd55bffd1b77fd82e5
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>