kubectl: upgrade /usr/bin/kubectl to 1.15.5
Tested this locally on the toolsbeta bastion.  Because the old cluster
requires our downloaded 1.4.12 kubectl, there is no use in having the
1.10 kubectl that kubernetes-client package installs.  1.10 kubectl causes
errors and warnings when interacting with the new cluster and is questionable
when working with the old. The only thing that uses /usr/bin/kubectl is
"webservice shell". Interestingly, 1.15.5 works great with the old cluster
as long as you are just doing "webservice shell", so this seems like a win.

Bug: T214513
Bug: T236202
Change-Id: If61477c74f37d685e2a5149e9d3af8d09ea08d1a