md: Globally set lower sync limits
Why:
kubernetes hosts were bitten by the same issue that etcd hosts have been
bit in T162013, namely that the RAID resync speeds are high enough to
cause IO latencies issues and thus facter failing. Since this has served
us fine up to now, there is good reason to generalize it across the
fleet. Note that one of the extra goodies coming with it is that the
mdadm cron is being moved to only be executed during a weekday,
minimizing the chances an issue will show up during a weekend.

What:
Move the code from the etcd profile to the raid/ module to enable it
globally

Bug: T237197
Change-Id: Id1d25cb8c59856fde540e64a3b8efb4edf917cf3