phabricator: monitor the PHD process if PHD is set to running in Hiera
The PHD process can be set to running/stopped individually on each
Phabricator server and we want to enable in on the codfw server as well.

So making it dependent on the "active_server" setting is not appropriate
anymore. We need to make it dependent on the actual Hiera setting for
the service.

HTTPS monitoring though should stay as it is and only exist once because
it is attached to a virtual server "phabricator".

Bug: T232883
Change-Id: I257f6e8a01945baee3d4bb27791188b2e038ad22