systemd::timer::job: fix bug re: On(In)?ActiveUnitSec
Without this, we will happily install the timer unit and the service unit,
and activate the timer unit -- but as far as systemd is concerned, that
doesn't *do* anything, because the service unit has never been activated
nor deactivated (and those times are what the OnActiveUnitSec and
OnInactiveUnitSec timer intervals care about).

So we need to start the service unit once, to set the timer in motion.

Change-Id: I1b590a75efb5133d5fd606664e765f68129f8fc0