role::dumps::distribution::server: add kerberos
In this change we add basic Kerberos client packages and
settings, plus a keytab for the 'dumpsgen' user, that will
be used once Kerberos will be enabled on Hadoop to
properly authenticate and fetch data.

Downside: the keytab is created for each host, so adding/moving
this role to another server will mean generating a new keytab.

Bug: T234229
Change-Id: I30735f9f2835a313edfc2155facc72948f139731