ATS: add sysconfdir to ReadWritePaths if ocsp is enabled
The unit trafficserver-tls.service runs update-ocsp-all as part of
ExecStartPre if ocsp is enabled. Such program needs to be able to write
under 'sysconfdir' or it fails as follows:

update-ocsp-all[25747]: touch: cannot touch
'/srv/trafficserver/tls/etc/ssl_multicert.config': Read-only file system

If ocsp is enabled, add 'sysconfdir' to the list of directories to be
mounted read-write.

Change-Id: I7277108f87af53bc27f75a097b594cfa19d61858