cloud: refactor prometheus role
Refactor role::prometheus::tools into a more modern layout. Since this role had
other clients outside Toolforge (the monitoring CloudVPS project), introduce a
couple of new roles:

 * role::wmcs::prometheus
 * role::wmcs::toolforge::prometheus

Both include the new refactored profile::toolforge::prometheus class, and the
toolforge specific role also includes the toolforge base classes, because it
was being used as a secondary role anyway.

While at it, refresh the prometheus blackbox role, which is now a profile as
well.

When this patch is merged, the following changes need to happen in CloudVPS:

* instances in the 'monitoring' project need to stop using
  role::prometheus::tools and start using role::wmcs::prometheus
* instances in the 'tools' project need to stop using role::prometheus::tools +
  role::toollabs::base and start using role::wmcs::toolforge::prometheus

No hiera updates should be required. The puppet catalog diff should be minimal,
other than new class names. In the case of Toolforge, we gain some additional
setups from the ::base profile which are a good addition anyway.

Bug: T238096
Change-Id: I4b4015a75d8a91905e80a0e3159bb991054d8be4
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>