mariadb-client: Install 10.4 on buster, unblock os upgrade
MariaDB 10.4 has recently been uploaded to the wikimedia repository.
Install it on cumin hosts and non-root clients on buster.

We may change it in the future to wmf-percona-server80-client or
wmf-mysql80-client for better compatiblity, but neither of those
packages exist yet.

This is mainly to avoid blocking buster upgrade of cumin
hosts; a definitive setup will be decided at a later time, based
on mysql production upgrade path.

Bug: T193224
Change-Id: Ib27970011bf0ac033c3e1b883fbe5619e36949b6