presto::server: add the presto user to the catalog
The presto user/group are created by the presto-server
deb package, but our code to deploy kerberos keytabs requires
the user to be defined in the puppet catalog to work properly.
This is similar to what we have done for hive/oozie/etc.. to
have a convenient way to require a user in puppet, even if it
is created by a deb package.

Change-Id: Ib8b8e32d2c34478f5ae66b445416f2e560de2164