admin: add analytics-privatedata system user
Add a system user called analytics-privatedata
to the analytics-privatedata-users POSIX group.
This user will be useful when Kerberos will be enabled
in the Analytics Hadoop cluster, since researchers/analysts/etc..
will need a way to run authenticated recurrent jobs (like crons)
without the need to kinit manually every 24h.

This is a interim compromise to avoid impacting users with Kerberos too much,
that may eventually be deprecated in favor of something better and more granular.

The idea is the following:

- a user authenticates with kinit (and password)
  when logging in into a stat/notebook host.
- a user in analytics-privatedata-users will also be able to sudo as analytics-privatedata
  to read a Kerberos keytab (owned by the user) and kinit/autheticate with it
  only on some stat/notebook hosts.

Bug: T238306
Change-Id: Iee3eeb1e00a24ff90ee60e5f8124a55ddeac74f3