bacula: Fix calculation of success rate function on bacula check
The rate calculation was added in a rush (but not used). Make it
actually work so it can be calculated weekly for metrics reporting.

This will require further work at a later time, for now just making
the function be right. This requires unit testing to prevent
future issues like this.

Change-Id: Ic6aee0acfc39f6217ccd9bb41961341b056d0065