vcl: move XWD pass logic to wm_common
Part of the implementation of X-Wikimedia-Debug consists in skipping the
cache for debug traffic to allow developers to test the behavior of
origins. Currently, such logic is defined in the VCL for cache_text
only. Move it to wm_common to pass XWD traffic for misc websites such as
noc.wikimedia.org as well.

Bug: T233768
Change-Id: Ia6e27ffcbc2601c5eee8348aa910165ad988e76e