puppet_compiler: ensure we dont reset the permissions of the repo
Currently there is a race condition which occures if a puppet run
happens during a PCC run.  At this point puppet will reset all the file
permissions under /srv/jenkins-workspace/puppet-compiler to 644.  This
prevents the config_version script from executing successfully

This CR updates the permissions so that only the owner is updated
recursivly

Change-Id: I97a953ab421df44cddfbde7ed2d69cd6195f1619