toolforge: proxy: adjust setup for the new k8s cluster
The toolforge front proxy runs nginx+lua to fetch info from redis to know where
to redirect requests. That is still valid for the webservices grid and the
legacy k8s. For the new k8s cluster we want to add a fall-through proxy
redirect to the haproxy in front of the ingress.

Don't merge this change until we are ready to go live with the new k8s cluster.
Among other things, this change will change how 404 is done in the legacy
setup. Before this patch, the LUA snippet would return the admin tool as
default option in case the requested tool wasn't available in the redis
database. With this patch, there is no such logic in LUA, and in case no option
is available in redis, nginx will proxy traffic to the new k8s cluster. Then,
this new k8s cluster should handle 404 situations etc.

Additional docs:
 https://wikitech.wikimedia.org/wiki/Portal:Toolforge/Admin/Networking_and_ingress

Bug: T234037
Change-Id: I895da3cd6c0763420d4d71d0357cf4d247148d9b
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>