toolforge: prometheus: enable scraping for the new k8s cluster
Enable prometheus service discovery and scraping for the new Toolforge
kubernetes cluster.

This includes the required config for kubernetes as well. We need some specific
RBAC bits for a new user account 'prometheus' with permissions to do the
discovery and scraping. This user is authenticated by means of TLS clients
certs, which should be generated in the k8s with the utility script we have.

The TLS cert key should be present in the private repo.

Bug: T237643
Change-Id: Ibe7fb223c08749b6520ce3aaa24f82b5bcb54ed6
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>