TLS Analytics: make parsing more robust
We've seen data (luckily only for four whole requests) in turnilo with
the "TLS Version" field set to "H2=1", which indicates something went
horribly wrong with either garbled header output by ats-tls, or some
kind of mis-parsing of the header in varnish-fe.  They were all for
beacon POSTs.

Based on the current code and regexes, the most-likely case was that the
contents of the raw header field "SSL=X" did not have an X matching
"TLSv1(\.[123])?", but we don't know what it actually had in this case.

This broadens the three non-binary field definitions' regexen to
[^; ]+ and uses this consistently, and adds an overall sanity
check regex of the whole header to ensure that misparses are
clearly indicated as such.

Change-Id: Id62ce6dd03e93e5ec79d0faeeb3b5e3ae6d92035