interface: Add a config option to avoid CPU 0 in RPS.
In a recent incident, it was discovered that when network queues are
spread across all CPUs, the result is increased load on CPU 0 -- since
other work uses that CPU by default. This patch adds a config option to
skip CPU 0 in RPS, spreading across all other CPUs instead.

The default behavior is unchanged, and this patch doesn't turn the flag
on anywhere.

Bug: T236208
Change-Id: I3306eb8a5ffd729a8a4546bd369168702586b5c0