prometheus-bacula-exporter: Setup bacula collection on prometheus
After I8dc98d64d4a68b9, start collecting metrics from the bacula
exporter from the prometheus hosts, by setting up the appropiate
job on prometheus servers.

For now, this should be a single host on eqiad, but that could
change in the future (because of a switchover, or setting up
multiple backup directors in parallel).

Getting bacula statistics can take a long time, so by default,
only scrape every hour, with a timeout of 1 minute.

Bug: T234900
Change-Id: Iff88d3e80bfa1ce7a8f2efc47ae78597b4d09a48