ATS: explicitly skip the cache instead of hiding CC
To skip the cache, explicitly disable TS_LUA_CONFIG_HTTP_CACHE_HTTP for
the current transaction instead of hiding Cache-Control from the ATS
state machine.

This will make responses for which we explicitly skip the cache show up
as "pass" in X-Cache-Int instead of "miss". Other than being the right
thing to do, this should help finding out why we are getting a lower
real hitrate on ats-be compared to varnish-be.

Bug: T238494
Bug: T227432
Change-Id: I278ff8b117f9cf42e55cbf181ddee28d8a08a4e8