puppet-export-facts: cache the cacert value
A previous update made change this code so it would get the cacert from
the value configured is puppets localcacert.  however the code is such
that this value is resolved for every host.  This updates the code we we
only resolve the file path once and cache it for all other requests

Change-Id: I0b4c2dfcad28f5ea3227f721e8bc37e9cd1f1c29