webperf: Remove xhgui profile from webperf::profiling_tools role
The xhgui profile is still an incomplete work in progress, and not
applied in production. This was added here in the past for testing
in Beta Cluster ahead of prod.

We've made some progress since then but it is still incomplete.

Meanwhile, the plan for prod has changed. XHGui will not be installed
on the webperf1002 (profiling_tools) host, but instead on a new
xhgui1001 host.

In order for beta's webperf13 to be like webperf1002 in prod,
remove the xhgui profile from it (again).

Bug: T180761
Bug: T238788
Change-Id: Idccfdf0968f5b28192c77fb90da0d6eef5903054