phab: avoid failing to load trusted_proxies in codfw after ATS migration
Since I14a865c3513acbb09 this lookup broke in codfw because there are
no more cache text nodes in codfw.

Leading to failed puppet runs on phab2001 with:

"The value '' cannot be converted to Numeric. "

Add a special case for codfw until we can switch it for eqiad as well.

Change-Id: I322f555acbdfb87146a1980f8fe7445b24dbe6f3