prometheus-bacula-exporter: Add expected fresness metrics
Two additional metrics per job are added:
* bacula_job_expected_freshness_seconds
* bacula_job_expected_full_freshness_seconds

Which measure the maximum amount of seconds between backups, and
full backups respectivelly. This will allow to indicate or alert
if a job has become stale (as this configurable options is needed,
in addition to the date of the last backups).

This will allow to have a check similar to that of icinga, but
with only prometheus metrics. Even if an alert is not setup, this
will be very helpful for debugging the icinga alerts.

Also fix misc bugs/style pointed on review.

Bug: T234900
Change-Id: Iab31fdca7ff61f7349fac639acb9aeae848f5c41