prometheus-bacula-exporter: Add total and successful backup counts
As requested, provide additional metrics about the total number of
backups attempted and the total number of backups successful in
the last month, week and day to be able to calculate the success
ratio of backups on a grafana dashboard.

Redo how freshness and backup count is counted to prevent too
much overhead. Additional improvements will require run bconsole
in parallel.

Bug: T234900
Change-Id: I8a490321c6ca0905177f2607042fc577ccbbe851