authdns: refactor role/profile/hieradata bits
This moves a bunch of role-level things down to the profile level
where they more-properly belong, and moves some manifest-level
data out to hieradata, imperfectly.  Other than dropping the
unused "authdns::testns" role, this is meant to be a refactor with
no real functional change for the authdns::server boxes.

It's also just an intermediate step, after which we need to
rationalize this and the dnsrecursor profiles so that the two can
be brought together in a combined role.

Bonus points: kills 10 puppet-lint style violations without
introducing any new ones.

Bug: T98006
Change-Id: Ie52939234930deee3ecda5ec391d9ba42fa2346a