xhgui: install unpuppetized apache php module
On the existing server, tungsten, libapache2-mod-php5 is installed
but it's not (anymore) puppetized as part of previous refactoring
from apache to httpd module.

Support installing it on newer distro versions as well.

Also, php-fpm is not actually installed here, so fix the path
to the php.ini files.

Switching to php-fpm would be a separate thing, emulating tungsten
only for now.

Bug: T238788
Change-Id: I0bc7f4d59cf9331d0a7570f32d6fdcfab3b421c1