browsersec: cover bot traffic better
Some non-human UAs are fetching APIs and such exclusively and may
not ever trip the warning if it's filtered down to just /wiki/
URIs, but there are many special-case URIs we don't want to impact
for the common human UAs.

This adds a user-agent filter so that we trap more-broadly when
the UA string is empty or matches /[Bb]ot/, which covers most of
these cases in practice, especially for the bad-tls requests.

Bug: T238038
Change-Id: I05cf431fd83c70a00b9ed96a4399c838363e4dfd