nova: add support for the 'placement' api service
The placement server is a new nova API that tracks resource
inventories.  It's more detail than we need for our setup but
in future OpenStack releases nova-compute won't run without it.

This also introduces a 'region' hiera setting because the placement
config requires specifying a region in the API config.

Bug: T239161
Change-Id: Ie2be4906cb0cda7cc8fd427a30574bf159c8bae0