ATS: log Cache-Control and Set-Cookie responses from origin
Add two origin server response headers: Set-Cookie (truncated to 16
characters) and Cache-Control.

Stop logging the backend request URL as it is often a fairly long field
and it is almost always the same as the request URL but using the origin
hostname. We can add it again in the future if necessary to debug issues
with specific URL rewrites.

Bug: T238494
Change-Id: I8aaf731467905b7f4a0ac2a2ff5ba52d02c3faf2