Add and use check_dns_query
The existing 'check_dns' is kind of awful: it's a C program which
just parses arguments to execute the even more-awful C program
'nslookup', and then parses its output for nagios.  More
importantly than all of that, it doesn't support using a
non-default DNS server port (and we'll need that for the upcoming
blended DNS role), and I'm honestly not even confident it parses
nslookup output as we expect in all cases.

So here's a Perl-based one that can be loaded as an efficient ePN
for Nagios (otherwise I'd just use python, probably?), checks all
the important things (including both transports, zero-answers
checks, header-vs-data mismatches, etc), and supports non-default
ports.

Bug: T98006
Change-Id: I37a493fa629a0f47dfd8b6b650162b0bcb443bfc