ATS: enable reload for global Lua script
Try again to allow reloading global Lua scripts as we did in a827ea6, but this
time without using __reload__ that clearly does not do what we think as
documented here: https://phabricator.wikimedia.org/T233274#5691926

Passing configuration as argtb parameters to the global Lua scripts only works
at startup, which is when __init__(argtb) is called, and not upon reload. Get
the configuration by reading it from a file every time the script is loaded
instead.

This reverts commit b2eb9ceaabb9465d186c1ef7039f0fcbbd853b93.

Bug: T233274
Change-Id: I08d924d49db6850c1a519dacfea501d3ceecc04b