authdns: configure explicit service/monitor addrs
Prior to this commit, the authdns config was listening on the ANY
address, which sufficed both for all of the loopback-defined
official service IPs as well as the host's own IP for per-server
healthchecks and monitoring.

For a blended role we can't have any daemon listening to port 53
on the ANY address or it will block any port 53 listeners of the
other daemon.

So the plan here is to leave recursive DNS as-is, which already
listens on port 53 of its explicit service addresses as well as
the host-level addresses, and change authdns to a new model:

* Authdns service addresses are explicitly listened on separately
* Port 5353 of the ANY addresses are listened to for monitoring
  purposes only, so that various health/monitoring services can
  still execute per-server checks against them.
* Temporarily, this commit also defines a monitor-only listener on
  port 53 of the host-level IPv4 and IPv6 explicitly (not via
  ANY).  This will smooth the icinga-level transition for the
  healthchecks of the legacy authdns servers and will be removed
  before we get to the blended role config.

(Also, while here, moved the conftool_prefix hieradata lookup out
of the authdns module and up to the consuming profile where it
more-properly belongs).

Bug: T98006
Change-Id: Ib01bb77a26ea6f0c4e3276dc1643c6a1ecdfdbc0