Introduce systemd cgroup memory limits for stat1004
The analytics client hosts like stats and notebooks are used to crunch
data coming from Hadoop, and sometimes improper use of resources
lead to out of memory events that 1) trigger alerts 2) disrupt any
other job running on the host.

The idea is to apply a generous memory-based constraint to every user
as starter, to be tuned later on in case not sufficient.

stat1004 is the first host to get the new setting.

Bug: T212824
Change-Id: I1cf75672876d2e19a44eb31940ee45103d230d1c