authdns: define gdnsd user/group explicitly
Ever since we deployed DNS Cookies, there has been a
chicken-and-egg scenario that requires 2x agent runs in a row to
bring up a new authdns server.  This is because all config is
deployed before installing the package (to ensure that if the
package starts the service and gets live prod DNS traffic routed
to it, it will serve appropriate answers), but the cookie config
has to be owned by the daemon user which is created by the package
install.

This fixes the situation by having puppet create the user.

Change-Id: I6e134b6af7e3beeccd7170080c07d4596a879b1c