dnsbox: set up rec/auth machine-local dependency
This ensures that gdnsd must be running for pdns-recursor to be
running.  Otherwise the pdns-recursor would fail at resolving our
own names, because our authdns IPs are defined on the loopback and
there's no gdnsd to answer on them.  gdnsd almost never needs a
restart anyways, but this helps in those rare cases and/or on
reboot.

Bug: T98006
Change-Id: I0538456ada4561a96922b6ba4aeb896b38a81e9b