production-m5.sql: Add access to nova_cell0 for nova user
Access from cloudcontrol1003 (208.80.154.23) and cloudcontrol1004
(208.80.154.132) to the new nova_cell0 with the same grants as the
access to nova database.

Also add the already existing grant to nova database which wasn't
reflected on this file before.
Right now, on the database, the grant for nova database is too open, it
is "%", we should limit it to the cloudcontrol hosts.
Once these new grants are deployed, we should be able to remove the
existing 'nova'@'%' one.

After this patch is merged, the grants need to be added to the database
manually

Bug: T239170
Change-Id: I428eb49076fb954a8bfe48ef2f92053a79581bfe