toolforge-k8s: simplify calico upgrades and distribute calicoctl
This includes changing the k8s version to the current cluster default.

Using kubectl apply -f calico.yaml should update the deployments when
the version is adjusted in hiera, using this since we are already
a few patch versions behind with some important crash fixes in there. This
should never be done to upgrade major versions, etc. Just the patch releases.
Moving from 3.8 to 3.9 should get a new template with some serious comparison
in case there are significant differences. This is probably all that is needed
for 3.8.0 to 3.8.4 (currently needed), though.

Adding calicoctl will allow us to do some more interesting things with
calico with full validation. You can interact with calico via the CRD directly
with kubectl, but it won't be validated. Supported deployment methods
include a container, pod or binary, so this will pull the binary from
the Github with a sha256 check in puppet.

Change-Id: I8d1e24ae2c14014ef76995440913b087f6a7fc8d