mediawiki: Avoid unsafe ob_start inside php7-fatal-error.php
Fatal errors can happen at any point during the PHP process
for MediaWiki, including when output buffers (ob) are active.
And these output buffers are not reset per se (as seen in T236832),
and/or there may be built-in buffers we don't control.

Avoid using the PHP's built in trace formatter, which always
prints to the user and thus we needed to capture that.
Instead use debug_backtrace() directly and format its result
in a similar way ourselves.

This matches what we do in MediaWiki core, which presumably does
so for the same reason.

Also remove the stripping of the first frame. While this might
be useless to include in most/all cases, it is hard to verify and
reason about. Given it is expected to include mutiple frames
of which some may be redundant, it seems entirely fine to just
include always at little to no inconvenience.

Bug: T236832
Change-Id: I2b3d19df5a5c544fac0201a20c804f6022663c5e