ATS: re-use origin server connections for matching IPs
The default setting, "both" means that connections are re-used if both
the IP address and fully qualified domain name match.

Set the value to "ip" to only check the IP. Apply the setting only to
mw, api and restbase origins which are known to properly support virtual
hosts on reused connections.

Bug: T238494
Change-Id: If39e73ca532a2cbc040d33599d237f8778ab7237