prometheus: scrape calico felix agent
Why:
Having calico felix metrics in prometheus for all kubernetes nodes could
prove beneficial when debugging kubernetes issues.
We put this in k8s prometheus and not in ops just in order to have those
in the same prometheus instances for easy of configuration of dashboards

What:
Add prometheus configuration for populating calico-felix endpoints.

Change-Id: I24b52790e2dcb67de0b91d1b1d4570d1a43f1d5f