bacula: Rename schedule to monthlys., split schedule & jobdefaults
schedule was probably the only schedule setup. Rename it with
the same pattern as the other defined schedules, so we have
monthly (default), weekly and hourly schedules.

Split schedule definition and job definition. Jobs with a
different pool are going to be setup, but we cannot duplicate
jobdefaults definitions. Define the schedule once, and then
define as many jobdefinitions as needed using a common schedule.

Bug: T238048
Change-Id: Ie081e459b35b787a837424b31ad87d8d2e374b78