bacula: Increase expected freshness of monthly full backups
Because how days of the week work, 5 weeks could theoretically pass
between 2 "first day of the week" monthly backups happens.
For example, if the backup happens the 1st of a 30-day month on
a monday, the next one won't happen until 7 days in a week * 5 week=
35 days later, on the 7th of the next month.

Add the original additional 24 hours as a buffer, for a total
of 36 days max between montly backups, even if in most cases only 4
weeks will separate them (28 days).

This will fix the currently missleading alert on icinga/prometheus.

Bug: T234900
Change-Id: I8b3d8978337420e283a672401156e7da8fdfad48