bacula: Setup new pool for databases as well as its configuration
The "databases" pool is a new pool that can be used, with dedicated
resources. For now, set it up only on the primary dc until a
an offsite policy is decided.

Aside from making the pool available, without modifing the
"production" one, create the jobdefaults configuration to make
possible its usage.

Bug: T238048
Change-Id: Ibb35ef34e19be49486afbbd3875e6dd455a396ce