profile::refinery::job::spark_job: allow to pass a keytab
What:
Add an option to our spark job definition in puppet to pass
a keytab to Yarn's distributed cache. Enable the new feature
only for Spark Refine jobs for the moment.

Why:
In Refinery's DataFrameToHive we issue CREATE/ALTER sql statement to Hive if needed.
Spark is not aware of this code and by default it retrieves Delegation Tokens
only for HDFS/Hive-Metastore/HBase. When the JDBC connection to the Hive Server 2
is established (with Kerberos enabled), some credentials will be needed to be able
to login as $user. These credentials are explicitly provided as keytab, that is copied
(securely) by Yarn to its distributed cache.

Bug: T228291
Change-Id: I5f5bf90895cbda858f09f6ed85efc46ebf1e021f