Drop grafana/jessie from reprepro sync
In jessie there's a common thirdparty component. We no longer use Grafana
on jessie (production uses Buster and labmon is being reimaged to Buster
as well). This avoids spurious syncs of the grafana deb everything we
sync up thirdparty/jessie-wikimedia (e.g. to pull in new Jenkins packages
for CI)

Change-Id: I4210966bb2a112a6e8243de7659ebc64874dc1e3