ATS: add milestone timing logs, update TTFetchHeaders
Be more specific in milestone timing logs: instead of considering
TTFetchHeaders the time between session start (client accept) and the
end of origin server request headers parsing, make it the time between
the first origin server connection attempt (or shared session attached
in case of connection reuse) and when the first byte is received from
the origin. This way we do not include things like cache lookup time in
TTFetchHeaders.

Add the following separate timings: ClientTTFB, namely the time between
accept and the first byte written to the client, CacheReadTime,
CacheWriteTime, and TotalSMTime which represents the whole time spent in
the ATS state machine from accept to the very end of the transaction.

Bug: T238494
Change-Id: I0c2ffcd49556f67e18157a31afcc4be48f0cf417