bacula: Increase production and Databases retention to 90 days
Bacula's retention was reduced to 30 days while it was on the
old infrastructure due to space concerns. The new infrastructure
has way more available space, having, after a month, 20.62 TB
available out of 35TB configured for production, and 35TB for
databases. Given there are, in addition, 30TB still unallocated
on the LVM, and that the production array will not have to handle
database backups anymore, we will be able to increase the
retention from 30 days to the originally intended 90 days.

Despite backups being an exemption of the 90 day data retention
policy https://meta.wikimedia.org/wiki/Data_retention_guidelines ,
keeping it at 90 days seems the right thing to do both technically
and policy-wise.

Bug: T238048
Change-Id: I99aa0d74ddc6252a30006e5b0b49b3b5841a2fc6