netbox: create netbox_frontend global variables
The netbox frontend is required in multiple places as such we should
make it a global parameter.

This CR also makes use of the alias interpolation function in hiera so
that all variables are locally  scoped in hiera.

PCC: https://puppet-compiler.wmflabs.org/compiler1001/19786/
The changes for puppetdb are related to the other CR's in this chain

Change-Id: Ifd1e06de80a1c97d4c9b6a5effe10005c442fcd9