bacula: Schedule hourly copies of production backups to the offsite pool
Copy jobs duplicating data from the production pool were configured,
but were not running because they were not scheduled.

Add a schedule to copy them every hour to backup2001. Final
copy schedule may be changed at a later time based on performance
and speed of the copy process.

Schedule the copy, not the migration! Followup to 0faa04b and its
reversion.

Bug: T238048
Change-Id: Ib80e17fd7662536c242d72002321a4dd3f92fd6b