mediawiki: Add reqId/file/line to php7-fatal-error.php's 'message' field
The 'message' field was previously left undefined at the source,
and back-filled during indexing with a Logstash filter.

Instead, create it explicitly at the source like we do for MediaWiki,
and also make its format look more similar (with reqId and file/line)
being part of the main message.

== Details

Currently a fatal can be logged upto three times to Logstash,
like so

  | _t         | type    | host   |
--|
1 | PHP Fatal error: Maximum execution… exceeded in …/UstringLibrary.php on…
  | …T18:37:24 | syslog  | mw1347 |
--|
2 | Maximum execution time of 180 seconds exceeded
  | …T18:37:24 |mediawiki| mw1347 |
--|
3 | [Xef5Ow…] /w… PHP Fatal Error from line 683 of …/UstringLibrary.php: Maxi…
  | …T18:37:24 |mediawiki| mw1347 |

Where:
* 1 comes from php's own stderr, logged as simple string with no context
  about the web request or MediaWiki. Not very useful and we don't have
  much influence on this.
* 2 comes from the php-wmerrors extension, logged by the php7-fatal-error.php
  script to syslog directly where we try to recover some minimal amount
  context where possible, and ensure it ends up in the right place
  where monitoring tools and humans query it from.
* 3 comes from MediaWiki run-time with full ideal context. This is mostly for
  non-fatal exceptions only, although for a small subset of fatals it is able
  to handle them as well.

This commit makes the Logstash filter redundant by setting the
'message', 'type' and 'channel' directly at the source in the JSON
packet, the same way MediaWiki and other services do already.

It also makes the message more similar to the MW one, and more useful by
including the request ID and exception file/line directly in that message
instead of only in the context fields. Also add 'normalised_message' so
that we don't regress in our ability to aggregate/combine similar errors
by frequency (this is currently filled in from the auto-created 'messsage').

This commit also migrates the Parsoid-PHP conditional from a hardcoded
hostname filter for 'wpt*' to the ENV-variable based method agreed
upon in T235899 and <https://gerrit.wikimedia.org/r/546448>, and
<https://gerrit.wikimedia.org/r/548944>.

Change-Id: I612ea594e04c6a89caec69ee0cf3d004f14f5949